+++ packages/router/src/lib/helpers/shared.ts
@@ -1,6 +1,13 @@
-import { getNtpTimeSeconds as _getNtpTimeSeconds, RequestContext, getChainData } from "@connext/nxtp-utils";
+import {
+  getNtpTimeSeconds as _getNtpTimeSeconds,
+  RequestContext,
+  getChainData,
+  multicall as _multicall,
+  Call,
+} from "@connext/nxtp-utils";
 import { BigNumber, utils } from "ethers";
 
+import { cachedPriceMap } from "../../bindings/prices";
 import { getDeployedChainIdsForGasFee } from "../../config";
 import { getContext } from "../../router";
 
@@ -104,12 +111,33 @@ export const calculateGasFeeInReceivingTokenForFulfill = async (
 };
 
 /**
- * Gets token price in usd from price oracle
+ * Gets token price in usd from cache first. If its not cached, gets price from price oracle.
  *
  * @param chainId The network identifier
  * @param assetId The asset address to get price for
  */
 export const getTokenPrice = async (
+  chainId: number,
+  assetId: string,
+  requestContext: RequestContext,
+): Promise<BigNumber> => {
+  const cachedPriceKey = chainId.toString().concat("-").concat(assetId);
+  const cachedTokenPrice = cachedPriceMap.get(cachedPriceKey);
+  const curTimeInSecs = await getNtpTimeSeconds();
+
+  // If it's been less than a minute since we retrieved token price, send the last update in token price.
+  if (cachedTokenPrice && cachedTokenPrice.timestamp <= curTimeInSecs + 60) {
+    return cachedTokenPrice.price;
+  }
+
+  // Gets token price from onchain.
+  const tokenPrice = await getTokenPriceFromOnChain(chainId, assetId, requestContext);
+  cachedPriceMap.set(cachedPriceKey, { timestamp: curTimeInSecs, price: tokenPrice });
+
+  return tokenPrice;
+};
+
+export const getTokenPriceFromOnChain = async (
   chainId: number,
   assetId: string,
   requestContext: RequestContext,
@@ -136,4 +164,18 @@ export const getGasPrice = async (chainId: number, requestContext: RequestContex
  */
 export const getChainIdForGasFee = (): number[] => {
   return getDeployedChainIdsForGasFee();
+};
+
+/**
+ * Runs multiple calls at a time, call data should be read methods. used to make it easier for sinon mocks to happen in test cases.
+ *
+ * @param abi - The ABI data of target contract
+ * @param calls - The call data what you want to read from contract
+ * @param multicallAddress - The address of multicall contract deployed to configured chain
+ * @param rpcUrl - The rpc endpoints what you want to call with
+ *
+ * @returns Array in ethers.BigNumber
+ */
+export const multicall = async (abi: any[], calls: Call[], multicallAddress: string, rpcUrl: string) => {
+  return await _multicall(abi, calls, multicallAddress, rpcUrl);
 };