+++ packages/utils/src/providers.ts
@@ -1,0 +1,5 @@
+import { ethers } from "ethers";
+
+export const getSimpleRPCPRovider = (rpcUrl: string) => {
+  return new ethers.providers.StaticJsonRpcProvider(rpcUrl);
+};