+++ CHANGELOG.md
@@ -5,7 +5,14 @@
 - [router] delegate fee calc to chainreader
 - [sdk] integrate chainreader
 - [txservice] introduced provider syncing
+
+## 0.0.93
+
+- [router] Improved caching and auction response time
+- [sdk, router] Add method to allow slippage in prepared amount
 - [txservice] removed TransactionServiceFailure error, replaced with specific error types
+- [subgraph] Add convenience fields for analytics subgraph
+- [subgraph] Redeploy analytics subgraphs
 
 ## 0.0.92